Well maybe don't always destroy history

+++ classes/class-plugin.php
@@ -439,8 +439,10 @@ class Plugin {
 
         $this->io->db->updateFormSubmissionsTable($form, $newFields, $deletedFields);
 
-        // The history of the form also has to go. Otherwise the field names stay reserved.
-        $this->io->db->destroyHistoryFields($form);
+        if ($deletedFields) {
+          // The history of the form also has to go. Otherwise the field names stay reserved.
+          $this->io->db->destroyHistoryFields($form);
+        }
       }
     } catch (Error $e) {
       $msg = $e->getMessage();
